requirements.txt
playwright==1.41.0
behave==1.2.6
allure-behave==2.13.2
allure-python-commons==2.13.2
